chore: rename error handling files related to auth
