+++ packages/cms-frontend/src/lib/utils/errors/auth.ts
